Fix `:type` depending on existing variables
